@@ -6,12 +6,12 @@ version '3.3.1'
 group = 'com.braintreepayments'
 
 android {
-  compileSdkVersion 27
-  buildToolsVersion '27.0.3'
+  compileSdkVersion 28
+  buildToolsVersion '28.0.3'
 
   defaultConfig {
     minSdkVersion 15
-    targetSdkVersion 27
+    targetSdkVersion 28
     versionCode 39
     versionName version
 
@@ -29,13 +29,14 @@ android {
   lintOptions {
     textReport true
     textOutput 'stdout'
+    abortOnError false
   }
 
   resourcePrefix 'bt_'
 }
 
 dependencies {
-  implementation 'com.android.support:design:27.0.0'
+  implementation 'com.google.android.material:material:1.0.0'
   testImplementation 'junit:junit:4.12'
   testImplementation 'org.robolectric:robolectric:3.3.2'
   testImplementation 'org.mockito:mockito-core:2.8.9'
